# gin/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "用于演示gin+swag的图书管理接口",
        "title": "图书管理API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/book/add": {
            "post": {
                "description": "添加一本新图书",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "图书管理"
                ],
                "summary": "添加图书",
                "parameters": [
                    {
                        "description": "图书信息",
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Book"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "图书添加成功，Data为添加的图书信息",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "400": {
                        "description": "参数错误或图书已存在",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    }
                }
            }
        },
        "/book/delete/{id}": {
            "delete": {
                "description": "根据ID删除图书",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "图书管理"
                ],
                "summary": "删除图书",
                "parameters": [
                    {
                        "type": "string",
                        "description": "图书ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "书籍删除成功",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "404": {
                        "description": "书籍不存在",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    }
                }
            }
        },
        "/book/search": {
            "get": {
                "description": "获取所有图书信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "图书管理"
                ],
                "summary": "查询所有图书",
                "responses": {
                    "200": {
                        "description": "书籍列表获取成功，Data为Book数组",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    }
                }
            }
        },
        "/book/update/{id}": {
            "put": {
                "description": "根据ID更新图书信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "图书管理"
                ],
                "summary": "更新图书",
                "parameters": [
                    {
                        "type": "string",
                        "description": "图书ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "图书信息",
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Book"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "图书更新成功，Data为更新后的图书信息",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "400": {
                        "description": "参数错误或ID不一致",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "404": {
                        "description": "图书未找到",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.Book": {
            "type": "object",
            "required": [
                "author",
                "id",
                "stock",
                "title"
            ],
            "properties": {
                "author": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "stock": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "main.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}